requirements.txt
python-telegram-bot==20.5
gspread==5.12.0
oauth2client==4.1.3
pandas==1.5.3
numpy==1.24.4
tornado==6.1

